finish add media youtube and review

+++ detail/src/main/java/com/ferdsapp/detail/data/model/movie_details/MovieDetailReviewAuthorDetails.kt
@@ -13,5 +13,5 @@ data class MovieDetailReviewAuthorDetails(
     val avatar_path: String?,
 
     @SerializedName("rating")
-    val rating: Int?
+    val rating: Float?
 )
